Fix stack corruption
Keep a variable registered as a benchmark parameter from going out of
scope so that the point to that variable remains valid during the
benchmark execution.

It isn't clear what this variable is intended for since it isn't
referenced after the benchmark completes.

This is just a quick fix to avoid stack corruption without addressing
the intended use of this variable.