Fix stack corruption by variable out of scope
An unused "dummy" variable provided to the benchmark so the benchmark
could write to it was going out of scope during registation, causing
the benchmark to write to it when the registration function stack frame
was no longer active.  Fixed by simply changing its storage class to
static to force it to persist.  The value returned by the benchmark is
not used in this case.